Fix Complaints About Not Supporting MupenLua
The default mupenlua download has a filename of mupen64_lua.exe. As this does not match the mupen64-lua executable name previously in STROOPs config file, STROOP will fail to pickup the default mupenlua download. This leads to lots of complaints about mupenlua not working in STROOP, as well as modified config files and older versions of STROOP being recommended. This addition of the default filename fixes these issues. The version with a hypen is left in for prior compatibility.